// frontend/src/shared/types/store.ts
export interface ConversationPatch {
  _id: string;
  type?: Conversation["type"];
  group?: Conversation["group"];
  participants?: Conversation["participants"];
  lastMessageAt?: string;
  seenBy?: Conversation["seenBy"] | string[];
  lastMessage?: LastMessage | null;
  unreadCounts?: Record<string, number>;
  blockInfo?: DirectBlockInfo;
  moveToTop?: boolean;
}